Apply new internal header/source structure on task.{c,h}
