+++ libraries/ah_base/include/ah/internal/uring/task.h
@@ -1,0 +1,12 @@
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// SPDX-License-Identifier: EPL-2.0
+
+#ifndef AH_INTERNAL_URING_TASK_H_
+#define AH_INTERNAL_URING_TASK_H_
+
+#define AH_I_TASK_PLATFORM_FIELDS struct ah_i_loop_evt* _evt;
+
+#endif
